requirements.txt
Flask==2.3.2
requests==2.31.0
fpdf2==2.7.4
cx_Oracle==8.3.0
numpy==1.24.3
opencv-python==4.7.0.72
python-dotenv==1.0.0